+++ src/lib.rs
@@ -6,6 +6,7 @@ mod utils;
 
 pub use crate::libs::coverage::*;
 pub use crate::libs::intspan::*;
+pub use crate::libs::overlap::*;
 pub use crate::libs::range::*;
 
 pub use crate::utils::*;