+++ src/libs/mod.rs
@@ -1,3 +1,4 @@
 pub mod coverage;
 pub mod intspan;
+pub mod overlap;
 pub mod range;